fix gh-106: AttributeError: 'Pinguino' object has no attribute 'hex_file'.
Adds a new `DoesntHaveHexFileError` that will be raised when there is no hex
file (because the program has not been compiled yet). This keeps the
case where there is no hex file as an error (because it is) but makes it easy
to catch and handle it.

Now, when that error happens while uploading from the IDE, a dialog will pop up
indicating the user to compile the file before uploading.